src/ecs/pyproject.toml
[tool.poetry]
name = "aws-solutions-architect-project"
version = "0.1.0"
description = "Project for the ITBA Cloud Architecture Final Assessment"
authors = ["nmema <nicolas.a.mema@gmail.com>"]
readme = "README.md"
packages = [{include = "aws_solutions_architect_project"}]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^20.0"
boto3 = "^1.26.57"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
